Fix import misnomer, lock down dependencies
There was an import statement looking for a "LevelDOWN" property on `module.exports`, but it had been changed in subsequent versions to "AbstractStreamLevelDOWN". This has been modified to look for the correct property and import it under the old name.

The reason the breaking change made it through was from a loosely defined dependency (abstract-stream). It was always installing the latest version, regardless of version/breaking changes. It has since been locked down to a specific minor version.

Fixes issue #6